refactor: move `ExporterStateInitializer` to upper level
