+++ zeebe/dynamic-config/src/main/java/io/camunda/zeebe/dynamic/config/ClusterConfigurationInitializer.java
@@ -42,10 +42,9 @@ import org.slf4j.LoggerFactory;
  * <li>When the local configuration is empty, a non-coordinating member waits until it receives a
  *     valid configuration from the coordinator via gossip. See {@link GossipInitializer}.
  * <li>After initialization, the configuration can be modified using {@link
- *     ClusterConfigurationModifier}. Currently only {@link
- *     io.camunda.zeebe.dynamic.config.ClusterConfigurationModifier.ExporterStateInitializer} is
- *     available. It overwrites the local member's state to keep it in sync with the statically
- *     configured exporters.
+ *     ClusterConfigurationModifier}. Currently only {@link ExporterStateInitializer} is available.
+ *     It overwrites the local member's state to keep it in sync with the statically configured
+ *     exporters.
  */
 public interface ClusterConfigurationInitializer {
   Logger LOG = LoggerFactory.getLogger(ClusterConfigurationInitializer.class);
